pyproject.toml
[project]
name = "timescale-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.5",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "uvicorn>=0.32.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
